Correcting the value of the constant used
In the potential formula, the original constant used is 8.9e-12. This is later scaled up to 8.9
The constant that should have been used is 8.9e9. This will later be scaled down to 8.9.
Since the final value is the same it will not affect the simulation in any way, but I'm changing it for the sake of accuracy.